Optional namespace (#276)
* feat: use current context namespace by default

Use namespace from current context for majority of bonfire commands,
making the --namespace optional.

The deploy command tries to use current namespace, and if it is not
reserved or not owned it reserves a new one.  This is to keep backwards
compatiblity.

This new behavior will not apply if bonfire is in "bot" mode so that automated tests in CI tools (e.g. Jenkins) remains the same.

* feat: add reserve option for deploy cmd

Adds --reserve option to deploy command to force a reservation of new
namespace.

Co-authored-by: Brandon Squizzato <bsquizza@redhat.com>