Allow fine-grained control over dependency removal
In some cases, uses may want to be selective about dependency removal.
For example, if some dependencies are being mocked during application
testing (e.g. with WireMock), the user might still wish to retain other
dependencies.

This patch changes the --no-remove-dependency and --remove-dependency
options to accept a slash after a component name followed by a
dependency.  If no slash follows, the component name, all dependencies
will be operated on.

Obviously this creates a resolution problem if the user provides
something like --remove-dependencies foo --no-remove-dependencies
foo/bar.  The code follows two rules: "specific beats general" and "keep
by default".  So in the example, the result would be the removal of all
dependencies except for foo/bar.  The resolution is handled through set
operations.  Or in the example --remove-dependencies component/b
--keep-dependencies component/c with an application that had the
dependencies a, b, and c, the result would be the retention of the
dependencies a and c.

This patch also adds substantial validation to catch cases that may be
technically resolvable but are more ambiguous.  For example,
--remove-dependencies foo --remove-dependencies foo/bar.  While this
case is resolvable without too much thought, I didn't want to introduce
additional complexity into what is already a fairly complex process.