Only add dependencies that are descendants of requested components (#505)
Suppose there is an app with components A, B, and C.  And with
dependencies of A', B', and C' respectively.  Prior to this patch,
bonfire had an issue where the final list of items from a `process`
command would be C, C', A', and B'.  The logic was such that if a
component was a dependency anywhere, it would be added.

The code that caused that bug was reverted, but that left another issue
where *only* listed components were added.  In that case, the final list
of items would be only C.

This patch changes the logic so that only C and C' will be in the final
result.